Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double stupeflip;
-    scanf("%d",&stupeflip);
-    printf("You entered: %.4d",stupeflip);
+    float stupeflip;
+    scanf("%f",&stupeflip);
+    printf("You entered: %.4f",stupeflip);
     return 0;
 }
